requirements.txt
# Production Requirements for Bww-AI-Assistant

# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy>=2.0.35
alembic>=1.13.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.8.0

# HTTP & API
requests>=2.32.0
python-multipart>=0.0.9
aiohttp>=3.9.0

# AI & Machine Learning
google-generativeai>=0.8.0

# System Monitoring
psutil>=5.9.0

# Utilities
python-dateutil>=2.9.0
pytz>=2024.1
rapidfuzz>=3.0.0
pyyaml>=6.0.1

# Development tools
pytest>=8.0.0
pytest-asyncio>=0.21.0

# Production Dependencies (Optional)
# gunicorn>=23.0.0  # Uncomment for production deployment
# psycopg2-binary>=2.9.9  # Uncomment for PostgreSQL support
# redis>=5.0.1  # Uncomment for caching support

# Monitoring & Logging (Optional)
# structlog>=24.1.0  # Uncomment for structured logging
# sentry-sdk[fastapi]>=2.15.0  # Uncomment for error tracking

# Security (Optional)
# cryptography>=43.0.0  # Uncomment for encryption
# passlib[bcrypt]>=1.7.4  # Uncomment for password hashing